app/requirements.txt
Django==6.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
requests==2.32.3
python-dotenv==1.0.0
django-cors-headers==4.3.1
djoser==2.2.2
django-simple-history==3.7.0
django-dump-load-utf8==0.0.4
django-codemirror2==0.2
django-jsoneditor==0.2.4
psycopg2-binary==2.9.9
gunicorn==21
django-filter==25.2
drf-spectacular==0.29.0
flake8==7.3.0
phonenumbers==9.0.29
django-phonenumber-field==8.4.0
celery==5.5.3
redis==6.1.0
django-redis==5.4.0
opensearch-py==2.8.0
channels==4.2.2
channels-redis==4.2.1
uvicorn==0.34.2
daphne==4.1.2
websockets==15.0.1